Move License from access.py to license.py.
